Pin bulma-phlex gem to main branch
This pins bulma-phlex to main. The released versions of the gem do not have the updated namespace. Once those updates are release, this will be updated. The two changes will need to occur together.

The development and test dependencies are in the Gemfile, not the gemspec. I do not have strong feelings on this, but followed the default Rubocop config.